Refactor tests, reduce size_t usage within chain/message types.

+++ include/bitcoin/bitcoin/chain/block.hpp
@@ -59,7 +59,7 @@ public:
 
     void reset();
 
-    size_t satoshi_size() const;
+    uint64_t satoshi_size() const;
 
     static hash_digest generate_merkle_root(
         const transaction_list& transactions);
